requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.1
django-redis==5.4.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
inflection==0.5.1
jmespath==1.0.1
packaging==23.2
Pillow==10.1.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
